Fix compile/linker errors in AstraSensors

+++ include/AstraSensors.h
@@ -96,7 +96,7 @@ void pullBMPData(Adafruit_BMP3XX &bmp, float (&bmp_data)[3]);
 
 float getBNOOrient(Adafruit_BNO055 &bno);
 
-void getPosition(SFE_UBLOX_GNSS &myGNSS, float (&gps_data)[3]);
+void getPosition(SFE_UBLOX_GNSS &myGNSS, double (&gps_data)[3]);
 
 String getUTC(SFE_UBLOX_GNSS &myGNSS);
